remove double code & some improments.
